requirements.txt
pandas>=2.2
scikit-learn>=1.4
joblib>=1.4
google-cloud-storage>=2.16
google-cloud-aiplatform>=1.50
python-dotenv>=1.0          